Revert combined rebates code, as data fetching got too messy
