+++ app/client/src/components/dashboard.tsx
@@ -145,11 +145,7 @@ export function Dashboard() {
     };
   }
 
-  if (
-    epaUserData.status !== "success" ||
-    csbData.status !== "success" ||
-    samEntities.status !== "success"
-  ) {
+  if (samEntities.status !== "success") {
     return <Loading />;
   }
 
@@ -181,7 +177,9 @@ export function Dashboard() {
       <div className="desktop:display-flex flex-justify border-bottom">
         <nav className="desktop:order-last mobile-lg:display-flex flex-align-center flex-justify-end">
           <p className="margin-bottom-1 margin-right-1">
-            <span>{epaUserData.data.mail}</span>
+            <span>
+              {epaUserData.status === "success" && epaUserData.data.mail}
+            </span>
           </p>
 
           <a
